% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvBayes}
\alias{cvBayes}
\title{Cross-Validation Without Doing Cross-Validation on Gibbs Sampling}
\usage{
cvBayes(Y, B, X, varE)
}
\arguments{
\item{Y}{is a phenotypic matrix with n rows and 1 column.}

\item{B}{is a matrix with the posterior distribution of markers effect.}

\item{X}{is a matrix with markers in columns and animals in rows.}

\item{varE}{is a vector with the posterior distribution of residual variance.}
}
\description{
Performs a Leave-One-Out Cross-Validation Without Doing Cross-Validation using Gibbs Sampling results.
}
\examples{
## Not to run ##

## cvBayes(Y=y, B=Effect, X=X, varE=BA_varE)

## End(Not run)

}
