% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvBGBLUP}
\alias{cvBGBLUP}
\title{Cross-Validation Without Doing Cross-Validation on Gibbs Sampling}
\usage{
cvBGBLUP(n, Y, g, K, Vg, Ve)
}
\arguments{
\item{n}{is the number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{Y}{is a phenotypic matrix with n rows and 1 column.}

\item{g}{is vector with the BLUP solution for the genetic values.}

\item{K}{is a relationship matrix with m rows and m columns.}

\item{Vg}{is the posterior mean of the genetic variance.}

\item{Ve}{is the posterior mean of the residual variance.}
}
\description{
Performs a Leave-One-Out Cross-Validation Without Doing Cross-Validation using Gibbs Sampling results.
}
\examples{
## Not to run ##

## cvBGBLUP(n=nSamp, Y=y, g=gHat, K=G, Vg=gVAR, Ve=eVAR)

## End(Not run)

}
