% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBLUP_CV.R
\name{GBLUP_CV}
\alias{GBLUP_CV}
\title{GBLUP and Leave-one-out Cross-validation}
\usage{
GBLUP_CV(formula, data, K, CV = TRUE, folds = NULL, weights = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame containing the variables in the model.}

\item{K}{Covariance matrix for random effects.}

\item{CV}{(logical) if TRUE the Leave-one-out Cross-validation is performed, default TRUE.}

\item{folds}{is a vector with validation sets used to perform k-folds cross-validation. Should be NULL or a numeric vector. If it is NULL, LOOCV will be performed.}

\item{weights}{is an optional vector to allow for heterogeneous error variance: \eqn{Var[\varepsilon_i] = R_i \sigma^2_e}. Should be NULL or a numeric vector. The length of weights must be equal to the number of individuals.}
}
\value{
A list with genetic and residual variances; a vector with BLUPs for random effects; a vector for BLUEs of fixed effects; the log-likelihood. AIC and BIC statistics. If CV is TRUE, a vector with BLUPs for random effects using LOOCV method.
}
\description{
This function solves univariate linear mixed models by likelihood methods. The optimization methods is based on Efficient Mixed Model Association (Kang et al. 2008).
}
\examples{
## Not to run ##

## GBLUP_CV(Phen ~ Effect, data=Data, K=G)

## End(Not run)

}
\references{
Hyun Min Kang, Noah A. Zaitlen, Claire M. Wade, Andrew Kirby, David Heckerman, Mark J. Daly and Eleazar Eskin, 2008. Efficient control of population structure in model organism association mapping. Genetics 178:1709-1723. doi:10.1534/genetics.107.080101.

Gianola D, Schon C-C, 2016. Cross-Validation Without Doing Cross-Validation in Genome-Enabled Prediction. G3: Genes|Genomes|Genetics. 6(10):3107-3128. doi:10.1534/g3.116.033381.
}
