% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kCV_DG}
\alias{kCV_DG}
\title{K-folds Cross-Validation Without Doing Cross-Validation}
\usage{
kCV_DG(Y, K, lambda, folds, g)
}
\arguments{
\item{Y}{is a phenotypic matrix with n rows and 1 column.}

\item{K}{is a relationship matrix with m rows and m columns.}

\item{lambda}{is the ratio bwtween residual and genetic variance.}

\item{folds}{is a vector with validation sets used to perform k-folds cross-validation.}

\item{g}{is the BLUP solution for the genetic values.}
}
\description{
Performs a K-folds Cross-Validation Without Doing Cross-Validation using a relationship matrix.
}
\examples{
## Not to run ##

## kCV_DG(Y, K, lambda, folds)

## End(Not run)

}
