% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LOOCV_DG}
\alias{LOOCV_DG}
\title{Cross-Validation Without Doing Cross-Validation}
\usage{
LOOCV_DG(y, K, lambda, g)
}
\arguments{
\item{y}{is a phenotypic matrix with n rows and 1 column.}

\item{K}{is a relationship matrix with m rows and m columns.}

\item{lambda}{is the ratio bwtween residual and genetic variance.}

\item{g}{is the BLUP solution for the genetic values.}
}
\description{
Performs a Leave-One-Out Cross-Validation Without Doing Cross-Validation using a relationship matrix.
}
\examples{
## Not to run ##

## LOOCV_DG(Y, G, lambda)

## End(Not run)

}
