% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LOOrrDG}
\alias{LOOrrDG}
\title{rrBLUP Cross-Validation Without Doing Cross-Validation}
\usage{
LOOrrDG(Y, X, B, lambda)
}
\arguments{
\item{Y}{is a phenotypic matrix with n rows and 1 column.}

\item{X}{is a matrix with markers in columns and animals in rows.}

\item{B}{is matrix with markers effect.}

\item{lambda}{is the ratio bwtween residual and genetic variance.}
}
\description{
Performs a Leave-One-Out Cross-Validation Without Doing Cross-Validation using a relationship matrix.
}
\examples{
## Not to run ##

## LOOrrDG(Y, X, B, lambda)

## End(Not run)

}
