% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RRBLUP}
\alias{RRBLUP}
\title{Ridge Regression BLUP}
\usage{
RRBLUP(y, X, M)
}
\arguments{
\item{y}{a matrix with n rows and 1 column}

\item{X}{a matrix with n rows and x columns}

\item{M}{a matrix with n rows and m columns}
}
\description{
Solves a univariate mixed model of form \eqn{y=X\beta+Mu+e}
}
\examples{
## Not to run ##

## RRBLUP(y, X, M)

## End(Not run)

}
